Add tests and move the functio to utils.ts
