+++ src/app/utils.ts
@@ -540,3 +540,66 @@ export function getLocalStoragePaginationValue(id?: string) {
 
   return undefined;
 }
+
+export const searchPairs = (query: string, pairsList: any): any => {
+  const searchQuery = query.trim().toLowerCase().replace(/\s+/g, " ");
+
+  const hasTypoTolerance = (source: string, target: string): boolean => {
+    const maxTyposAllowed = Math.floor(source.length / 4);
+
+    if (source.length > target.length + maxTyposAllowed) {
+      return false;
+    }
+
+    if (target.includes(source)) {
+      return true;
+    }
+
+    for (let i = 0; i <= target.length - source.length; i++) {
+      const substring = target.substring(i, i + source.length);
+      let mismatchCount = 0;
+      for (let j = 0; j < source.length; j++) {
+        if (source[j] !== substring[j]) {
+          mismatchCount++;
+        }
+      }
+      if (mismatchCount <= maxTyposAllowed) {
+        return true;
+      }
+    }
+
+    return false;
+  };
+
+  const preprocessPairName = (name: string): string =>
+    name.toLowerCase().replace(/\//g, " ");
+  const preprocessToken = (token: {
+    symbol: string;
+    name: string;
+  }): { symbol: string; name: string } => ({
+    symbol: token.symbol.toLowerCase(),
+    name: token.name.toLowerCase(),
+  });
+
+  return pairsList.filter((pair: any) => {
+    const pairName = preprocessPairName(pair.name);
+    const pairNameReversed = pairName.split(" ").reverse().join(" ");
+    const token1 = preprocessToken(pair.token1);
+    const token2 = preprocessToken(pair.token2);
+
+    const nameMatches = [
+      pairName,
+      pairNameReversed,
+      token1.symbol,
+      token2.symbol,
+      token1.name,
+      token2.name,
+    ];
+
+    return nameMatches.some(
+      (nameMatch) =>
+        nameMatch.includes(searchQuery) ||
+        hasTypoTolerance(searchQuery, nameMatch)
+    );
+  });
+};
